% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kde.R
\name{adjustClassifierParameter}
\alias{adjustClassifierParameter}
\title{Adjust Classification Tuning Parameter for fitted kdetrees Objects}
\usage{
adjustClassifierParameter(x, k, greedy = FALSE)
}
\arguments{
\item{x}{a fitted kdetrees object}

\item{k}{new classifier tuning parameter}

\item{greedy}{greedy outlier classification?}
}
\value{
a refitted kdetrees object
}
\description{
Takes a fitted kdetrees object and recalculates the outliers based on a new tuning parameter 'k'.
}
\author{
Grady Weyenberg
}

