% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kdetrees-package.R
\docType{data}
\name{apicomplexa}
\alias{apicomplexa}
\title{Apicomplexa gene trees sample data set.}
\format{a multiPhylo object with 268 phylogenetic trees}
\description{
Apicomplexa gene trees sample data set.
}
\examples{
kdetrees(apicomplexa, outgroup="Tt")
}

