% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.data.frame.kdetrees}
\alias{as.data.frame.kdetrees}
\title{Convert kdetrees object to data.frame}
\usage{
\method{as.data.frame}{kdetrees}(x, row.names, optional, trees = NULL, ...)
}
\arguments{
\item{x}{kdetrees object to be converted}

\item{row.names}{ignored}

\item{optional}{ignored}

\item{trees}{If given the original list of trees, will convert to
newick and add a column to the output}

\item{...}{unused}
}
\value{
a data.frame
}
\description{
Converts a kdetrees object to a data.frame suitable for saving as
output. It contains the density estimates for each tree, a Boolean
value indicating if the tree was selected as an outlier, and
optionally the newick string corresponding to the tree.
}
\examples{
result <- kdetrees(apicomplexa)
as.data.frame(result)
}
\author{
Grady Weyenberg
}

