% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhv.R
\name{bhv.consts}
\alias{bhv.consts}
\title{Find BHV constant lower bound}
\usage{
bhv.consts(trees, bw, hgm = TRUE)
}
\arguments{
\item{trees}{list of trees at which kernels will be centered}

\item{bw}{bandwidth for each kernel}

\item{hgm}{logical, if TRUE uses the HGM method, if FALSE uses a different method (default is TRUE)}
}
\value{
a vector of kernel integration constant lower bounds
}
\description{
Obtain a lower bound for the kernel integration constants
for centered on the trees with bandwidths bw.
}
\author{
Grady Weyenberg
}
