% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kde.R
\name{cutoff}
\alias{cutoff}
\title{Find cutoff value based on IQR}
\usage{
cutoff(x, k = 1.5)
}
\arguments{
\item{x}{score vector}

\item{k}{classifer tuning parameter}
}
\value{
cutoff value
}
\description{
Find cutoff value based on IQR
}
\author{
Grady Weyenberg
}

