% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expPoly.R
\name{dG.expPoly}
\alias{dG.expPoly}
\title{Pfaffian matrix for Exponential-polynomial distribution}
\usage{
dG.expPoly(th, G)
}
\arguments{
\item{th}{the vector theta of polynomial coeffiecients}

\item{G}{the value of the vector F(theta)}
}
\value{
the derivative matrix evaluated at point theta
}
\description{
This is the derivative matrix of the system. It should be arranged
as F_ij = dF(a)[j]/da[i]
}
\author{
Grady Weyenberg
}

