% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kde.R
\name{estimate}
\alias{estimate}
\title{estimate densities from kernel matrix}
\usage{
estimate(x, i = integer(), log = TRUE)
}
\arguments{
\item{x}{matrix of kernel contributions}

\item{i}{vector of columns to exclude from calculation}

\item{log}{log transform estimates?}
}
\value{
vector of density estimates for each tree
}
\description{
estimate densities from kernel matrix
}
\author{
Grady Weyenberg
}

