% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expPoly.R
\name{Gpartial}
\alias{Gpartial}
\title{Evaluate}
\usage{
Gpartial(th, G, m = 0)
}
\arguments{
\item{th}{Parameter vector}

\item{G}{solved F(th) vector}

\item{m}{order of derivative}
}
\value{
value
}
\description{
Evaluate the higher order derivatives of an solved
Exponential-Polynomial system (Hayakawa and Takemura 2014).
}
\author{
Grady Weyenberg
}

