% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expPoly.R
\name{hgm.expPoly}
\alias{hgm.expPoly}
\title{Normalizing constant for Exponential-Polynomial distribution}
\usage{
hgm.expPoly(theta, m = 0)
}
\arguments{
\item{theta}{vector of polynomial coefficients}

\item{m}{derivative order of F to evaluate}
}
\value{
Normalizing constant for the distribution
}
\description{
From Hayakawa and Takemura 2014
}
\author{
Grady Weyenberg
}

