% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhv.R
\name{internal.blens}
\alias{internal.blens}
\title{Extract vector of internal branch lengths}
\usage{
internal.blens(x)
}
\arguments{
\item{x}{a phylo object}
}
\value{
a vector with the internal edge lengths
}
\description{
Take a phylo object and extract a vector containing the internal
edge lengths, sorted by magnitude.
}
\author{
Grady Weyenberg
}
