% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdetrees-package.R
\docType{package}
\name{kdetrees-package}
\alias{kdetrees-package}
\title{kdetrees: An R Package for Kernel Density Estimation on Phylogenetic Trees}
\description{
The kdetrees package provides functions for kernel density estimation on phylogenetic trees.
}
