% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{load.trees}
\alias{load.trees}
\title{Read Newick trees from a set of files.}
\usage{
load.trees(paths, extension = ".tre", use.file.names = TRUE)
}
\arguments{
\item{paths}{If estension is NULL, a list of files containing
trees. Otherwise, the prefix for a wildcard expression.}

\item{extension}{If NULL, the paths are passed unaltered to
read.tree. Otherwise, a suffix (file extension) for the wildcard
expression.}

\item{use.file.names}{Set the tree names using the file
names. Will only work properly if each file contains a single
tree.}
}
\value{
A multiPhylo object.
}
\description{
Read a set of files containing individual trees and return a
multiPhylo object with all trees in the files. The set of files to
be read can either be passed explicitly as a character vector, or
through a combination of a set of directories and a file extension
which is used to construct a wildcard expression which will match
files with the extension you specify..
}
\details{
The behavior is determined by the extension parameter: If it is
NULL, then the paths supplied are passed unaltered to the
ape::read.tree function. If the parameter is not null, then the
paths are altered by calling Sys.glob with wildcard expression
"paths*extension".
}
\author{
Grady Weyenberg
}

