% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{sort.kdetrees}
\alias{sort.kdetrees}
\title{Reorder objects in a kdetrees object based on current density scores.}
\usage{
\method{sort}{kdetrees}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{kdetrees object}

\item{decreasing}{order of sorting}

\item{...}{additional parameters passed to \code{\link{order}}}
}
\value{
a new kdetrees object
}
\description{
Reorder objects in a kdetrees object based on current density scores.
}
\author{
Grady Weyenberg
}

