% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kde.R
\name{which.outliers}
\alias{which.outliers}
\title{Find outlier indices based on Tukey's IQR method.}
\usage{
which.outliers(km, k, greedy = TRUE)
}
\arguments{
\item{km}{the kernel matrix}

\item{k}{the tuning parameter}

\item{greedy}{greedy search?}
}
\value{
indices of outliers
}
\description{
Find outlier indices based on Tukey's IQR method.
}
\author{
Grady Weyenberg
}

