% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{apicomplexa}
\alias{apicomplexa}
\title{Apicomplexa gene trees sample data set.}
\format{a multiPhylo object with 268 phylogenetic trees}
\description{
Apicomplexa gene trees sample data set.
}
\examples{
kdetrees(apicomplexa)
}

