% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.matrix.multiPhylo}
\alias{as.matrix.multiPhylo}
\title{dissimilarity map tree vectorization}
\usage{
\method{as.matrix}{multiPhylo}(x, ...)
}
\arguments{
\item{x}{an ape::multiPhylo object.}

\item{...}{additional options for \code{ape::cophenetic.phylo}}
}
\value{
a row matrix of tree vectors
}
\description{
Dissimilarity maps convert trees to vectors using tip-to-tip path
lengths. Branch length information may be optionally discarded
(the default), resulting in vectors based solely on tree
topology.
}
\examples{
as.matrix(apicomplexa[1:5])
}
\author{
Grady Weyenberg
}

