% Generated by roxygen2 (4.0.0): do not edit by hand
\name{bhv.consts}
\alias{bhv.consts}
\title{Find BHV constant lower bound}
\usage{
bhv.consts(trees, bw)
}
\arguments{
\item{trees}{list of trees at which kernels will be centered}

\item{bw}{bandwidth for each kernel}
}
\value{
a vector of kernel integration constant lower bounds
}
\description{
Obtain a lower bound for the kernel integration constants
for centered on the trees with bandwidths bw.
}
\author{
Grady Weyenberg
}

