% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bhv.R, R/expPoly.R
\name{bhv.orthant.lb}
\alias{bhv.orthant.lb}
\title{Lower orthant bound for tree t with bw h.}
\usage{
bhv.orthant.lb(t, h)

bhv.orthant.lb(t, h)
}
\arguments{
\item{t}{}

\item{h}{}

\item{t}{}

\item{h}{}
}
\value{
lower bound volume

lower bound volume
}
\description{
Lower orthant bound for tree t with bw h.

Lower orthant bound for tree t with bw h.
}
\author{
Grady Weyenberg

Grady Weyenberg
}

