% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bhv.R
\name{bhv.orthant.lb}
\alias{bhv.orthant.lb}
\title{Lower orthant bound for tree t with bw h.}
\usage{
bhv.orthant.lb(t, h)
}
\arguments{
\item{t}{tree}

\item{h}{bandwidth}
}
\value{
lower bound volume
}
\description{
Lower orthant bound for tree t with bw h.
}
\author{
Grady Weyenberg
}

