\name{estimate}
\alias{estimate}
\title{estimate densities from kernel matrix}
\usage{
  estimate(x, i = integer())
}
\arguments{
  \item{x}{matrix of kernel contributions}

  \item{i}{vector of columns to exclude from calculation}
}
\value{
  vector of density estimates for each tree
}
\description{
  estimate densities from kernel matrix
}
\author{
  Grady Weyenberg
}

