% Generated by roxygen2 (4.0.0): do not edit by hand
\name{hist.kdetrees}
\alias{hist.kdetrees}
\title{Create a histogram of tree density estimates}
\usage{
\method{hist}{kdetrees}(x, ...)
}
\arguments{
\item{x}{kdetrees object to plot}

\item{...}{additional arguments passed to ggplot}
}
\value{
a ggplot object
}
\description{
Create a histogram of tree density estimates
}
\examples{
result <- kdetrees(apicomplexa)
hist(result)
}
\author{
Grady Weyenberg
}

