% Generated by roxygen2 (4.0.0): do not edit by hand
\name{internal.blens}
\alias{internal.blens}
\title{Extract vector of internal branch lengths}
\usage{
internal.blens(x)
}
\arguments{
\item{x}{a phylo object}
}
\value{
a vector with the internal edge lengths
}
\description{
Take a phylo object and extract a vector containing the internal
edge lengths, in no particular order.
}
\author{
Grady Weyenberg
}

