% Generated by roxygen2 (4.0.0): do not edit by hand
\name{kdetrees.complete}
\alias{kdetrees.complete}
\title{Complete kdetrees analysis convenience function}
\usage{
kdetrees.complete(infile, ..., treeoutfile = "outliers.tre",
  csvfile = "results.csv", plotfile = "plot.png", histfile = "hist.png")
}
\arguments{
\item{infile}{newick file with trees}

\item{...}{additional parameters for kdetrees}

\item{treeoutfile}{write outlier trees in newick format to this file}

\item{csvfile}{write density results to this file}

\item{plotfile}{print scatterplot of results to this file}

\item{histfile}{print histogram of density estimates to this file}
}
\value{
result of kdetrees call
}
\description{
Performs a complete kdetrees analysis, starting with reading trees
from a newick file on disk, and writing result files to the
working directory. Names and location of output files may be
controlled by optional arguments.
}
\author{
Grady Weyenberg
}

