% Generated by roxygen2 (4.0.0): do not edit by hand
\name{normkern}
\alias{normkern}
\title{Generalized Gaussian kernel}
\usage{
normkern(x, bw = 1, delta = 2L)
}
\arguments{
\item{x}{places to evaluate kernel}

\item{bw}{bandwidth values}

\item{delta}{shape parameter for kernel}
}
\value{
an object of the same type as x with the kernel evaluations
}
\description{
The un-normalized Gaussian kernel function: exp(-(abs(x/bw))^delta)/bw
}
\details{
The bandwidth parameter may be used in any way that makes sense in
the above R expression. In particular, it may be a single value,
for a constant bandwidth, or a vector, with each element
corresponding the bandwidth of the kernel to be placed at each
respective observation.
}
\author{
Grady Weyenberg
}

