% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bhv.R
\name{Northant}
\alias{Northant}
\title{Count semi-labeled rooted trees on n taxa}
\usage{
Northant(n)
}
\arguments{
\item{n}{number of taxa, must be at least 3}
}
\value{
integer number of topologies for a rooted phylogeny
}
\description{
Counts the number of topologies for rooted trees on n taxa. This
is the number of orthants in BHV tree space.
(NOTE) Candidate for optimization?
}
\author{
Grady Weyenberg
}

