% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Northant}
\alias{Northant}
\title{count semi-labeled rooted trees on n taxa}
\usage{
Northant(n)
}
\arguments{
\item{n}{number of taxa, must be at least 3}
}
\value{
integer number of topologies for a rooted phylogeny
}
\description{
counts the number of topologies for rooted trees on n taxa
(NOTE) Candidate for optimization.
}
\author{
Grady Weyenberg
}

