% Generated by roxygen2 (4.0.0): do not edit by hand
\name{plot.kdetrees}
\alias{plot.kdetrees}
\title{Plot the unnormalized density estimates for each tree.}
\usage{
\method{plot}{kdetrees}(x, ...)
}
\arguments{
\item{x}{kdetrees object to be plotted}

\item{...}{additional arguments passed to ggplot}
}
\value{
a ggplot object
}
\description{
Plot the unnormalized density estimates for each tree.
}
\examples{
result <- kdetrees(apicomplexa)
plot(result)
}
\author{
Grady Weyenberg
}

