\name{print.kdetrees}
\alias{print.kdetrees}
\title{Summarize a kdetrees object}
\usage{
  \method{print}{kdetrees} (x, ...)
}
\arguments{
  \item{x}{object to be printed}

  \item{...}{unused}
}
\value{
  invisible(x)
}
\description{
  Pretty-prints the results of a kdetrees ananlysis to
  console.
}
\author{
  Grady Weyenberg
}

