% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.kdetrees}
\alias{print.kdetrees}
\title{Summarize a kdetrees object in human-readable form.}
\usage{
\method{print}{kdetrees}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{...}{unused, required for generic compatability}
}
\value{
invisible(x)
}
\description{
Pretty-prints the results of a kdetrees ananlysis to console.
}
\author{
Grady Weyenberg
}

